New Feature

+++ src/main/kotlin/dev/vxrp/bot/ticket/TicketHandler.kt
@@ -1,37 +1,46 @@
 package dev.vxrp.bot.ticket
 
 import dev.minn.jda.ktx.coroutines.await
+import dev.minn.jda.ktx.messages.Embed
+import dev.minn.jda.ktx.messages.send
 import dev.vxrp.configuration.loaders.Config
+import dev.vxrp.configuration.loaders.Translation
 import dev.vxrp.database.sqlite.tables.TicketTable
+import dev.vxrp.util.color.ColorTool
 import net.dv8tion.jda.api.JDA
 import net.dv8tion.jda.api.entities.User
+import net.dv8tion.jda.api.entities.channel.concrete.ThreadChannel
+import net.dv8tion.jda.api.interactions.modals.ModalMapping
 import org.jetbrains.exposed.sql.SqlExpressionBuilder.eq
 import org.jetbrains.exposed.sql.deleteWhere
 import org.jetbrains.exposed.sql.insert
 import org.jetbrains.exposed.sql.selectAll
 import org.jetbrains.exposed.sql.transactions.transaction
 import org.slf4j.LoggerFactory
+import java.time.Instant
 import java.time.LocalDate
 
-class TicketHandler(val api: JDA, val config: Config, private val ticketType: TicketType, private val ticketStatus: TicketStatus, private val ticketCreator: String, private val ticketHandler: User?) {
+class TicketHandler(val api: JDA, val config: Config, val translation: Translation, private val ticketType: TicketType, private val ticketStatus: TicketStatus, private val ticketCreator: String, private val ticketHandler: User?, val modalId: String, val modalValue: MutableList<ModalMapping>) {
     private val logger = LoggerFactory.getLogger(TicketHandler::class.java)
     val settings = querySettings()
 
-    suspend fun createTicket() {
+    suspend fun createTicket(): ThreadChannel? {
         if (settings == null) {
             logger.error("Can't create ticket, settings are not loaded")
-            return
+            return null
         }
 
         val channel = api.getTextChannelById(settings.parentChannel)
         if (channel == null) {
             logger.error("Could not find specified parent channel for '{}' ticket", settings.name)
-            return
+            return null
         }
 
         val child = channel.createThreadChannel(settings.childRules.parentName.replace("%r%", retrieveSerial().toString()), true).await()
-        val creatorUser = api.retrieveUserById(ticketCreator).await()
-        child.sendMessage(creatorUser.asMention).await().delete().queue()
+        if (ticketCreator != "anonymous") {
+            val creatorUser = api.retrieveUserById(ticketCreator).await()
+            child.sendMessage(creatorUser.asMention).await().delete().queue()
+        }
         for (roleId in settings.roles) {
             val role = api.getRoleById(roleId)
 
@@ -41,7 +50,113 @@ class TicketHandler(val api: JDA, val config: Config, private val ticketType: Ti
             }
 
             child.sendMessage(role.asMention).await().delete().queue()
-            addToDatabase(child.id, LocalDate.now())
+        }
+
+        addToDatabase(child.id, LocalDate.now())
+        sendMessage(ticketType, child, ticketCreator, modalId, modalValue)
+        return child
+    }
+
+    private suspend fun sendMessage(type: TicketType, channel: ThreadChannel, userId: String, modalId: String, modalValues: MutableList<ModalMapping>) {
+        when(type) {
+            TicketType.GENERAL -> {
+                val embed = Embed {
+                    author {
+                        val user = api.retrieveUserById(userId).await()
+                        iconUrl = user.avatarUrl
+                        name = user.name
+                    }
+                    title = ColorTool().useCustomColorCodes(translation.support.embedTicketGeneralTitle.replace("%name%", channel.name))
+                    description = ColorTool().useCustomColorCodes(translation.support.embedTicketGeneralBody
+                        .replace("%issuer%", "<@$userId>")
+                        .replace("%subject%", modalValues[0].asString)
+                        .replace("%explanation%", modalValues[1].asString))
+                    timestamp = Instant.now()
+                }
+
+                channel.send("", listOf(embed)).queue()
+            }
+
+            TicketType.REPORT -> {
+                val embed = Embed {
+                    author {
+                        val user = api.retrieveUserById(userId).await()
+                        iconUrl = user.avatarUrl
+                        name = user.name
+                    }
+                    title = ColorTool().useCustomColorCodes(translation.support.embedTicketReportTitle.replace("%name%", channel.name))
+                    description = ColorTool().useCustomColorCodes(translation.support.embedTicketReportBody
+                        .replace("%issuer%", "<@$userId>")
+                        .replace("%reported%", "<@${modalId.split(":")[0]}>")
+                        .replace("%reason%", modalValues[0].asString)
+                        .replace("%proof%", modalValues[1].asString))
+                    timestamp = Instant.now()
+                }
+
+                channel.send("", listOf(embed)).queue()
+            }
+
+            TicketType.ERROR -> {
+                val embed = Embed {
+                    author {
+                        val user = api.retrieveUserById(userId).await()
+                        iconUrl = user.avatarUrl
+                        name = user.name
+                    }
+                    title = ColorTool().useCustomColorCodes(translation.support.embedTicketReportTitle.replace("%name%", channel.name))
+                    description = ColorTool().useCustomColorCodes(translation.support.embedTicketReportBody
+                        .replace("%issuer%", "<@$userId>")
+                        .replace("%problem%", modalValues[0].asString)
+                        .replace("%times%", modalValues[1].asString)
+                        .replace("%reproduce%", modalValues[2].asString)
+                        .replace("%additional%", modalValues[3].asString))
+                    timestamp = Instant.now()
+                }
+
+                channel.send("", listOf(embed)).queue()
+            }
+
+            TicketType.UNBAN -> {
+                val embed = Embed {
+                    author {
+                        val user = api.retrieveUserById(userId).await()
+                        iconUrl = user.avatarUrl
+                        name = user.name
+                    }
+                    title = ColorTool().useCustomColorCodes(translation.support.embedTicketReportTitle.replace("%name%", channel.name))
+                    description = ColorTool().useCustomColorCodes(translation.support.embedTicketReportBody
+                        .replace("%issuer%", "<@$userId>")
+                        .replace("%steamId%", modalValues[0].asString)
+                        .replace("%reason%", modalValues[1].asString))
+                    timestamp = Instant.now()
+                }
+
+                channel.send("", listOf(embed)).queue()
+            }
+
+            TicketType.COMPLAINT -> {
+                val embed = Embed {
+                    author {
+                        if (userId != "anonymous") {
+                            val user = api.retrieveUserById(userId).await()
+                            iconUrl = user.avatarUrl
+                            name = user.globalName
+                        } else {
+                            iconUrl = "https://www.pngarts.com/files/4/Anonymous-Mask-Transparent-Images.png"
+                            name = "Anonymous"
+                        }
+                    }
+                    title = ColorTool().useCustomColorCodes(translation.support.embedTicketReportTitle.replace("%name%", channel.name))
+                    description = ColorTool().useCustomColorCodes(translation.support.embedTicketReportBody
+                        .replace("%issuer%", "<@$userId>")
+                        .replace("%staff%", "<@${modalId.split(":")[0]}>")
+                        .replace("%proof%", modalValues[0].asString)
+                        .replace("%reason%", modalValues[1].asString))
+                    timestamp = Instant.now()
+                }
+
+                channel.send("", listOf(embed)).queue()
+            }
         }
     }
